fix all tslint errors
