+++ tslint.json
@@ -1,4 +1,5 @@
 {
+  "extends": "tslint-config-airbnb",
   "defaultSeverity": "error",
   "extends": ["tslint:recommended"],
   "jsRules": {
@@ -8,7 +9,8 @@
   "rules": {
     "quotemark": [true, "single", "avoid-escape", "avoid-template"],
     "object-literal-sort-keys": false,
-    "member-access": false
+    "member-access": false,
+    "interface-name": false
   },
   "rulesDirectory": []
 }
